refactor: moved the validation codes into its own file
