+++ src/controller/advertisement/mod.rs
@@ -1,1 +1,4 @@
 pub mod ad_controller;
+pub mod validation;
+
+pub use ad_controller::routes as advertisement_routes;
